// src/app/model/reservation-details.ts
export class ReservationDetails {

    

    reservation_id: string;
    chambre_id: string;
    NbrAdulte: string;
    NbrEnfant: string;
    NbrBebe: string;
    ageenfant: string;
    debut: string;
    fin: string;
    demandeSpeciale: string;
    etat: string;
    remarque: string;
    nbjours: string;


    



}